refactor: typed GraphNode/GraphEdge model + interactive dashboard
A5 - typed data model
- GraphNode: well-known attributes are now strongly-typed fields (Content,
  FilePath, StartLine, EndLine, ContentHash) instead of computed accessors over a
  Dictionary with PascalCase/camelCase dual lookups. Properties now holds only
  dynamic, parser-specific metadata.
- GraphEdge: drop the RelationType alias; Relationship is the single source of truth.
- Storage: ReadNode/Upsert map dedicated columns to/from the typed fields; the
  metadata blob round-trips Properties. Removed GetPropertyOrDbNull / GetIntPropertyOrDbNull;
  SerializeMetadata simplified.
- Updated all writers: Roslyn/JS/GraphQL/PHP/Markdown parsers, GraphIndexScanner,
  CrossTechLinker (reads node.FilePath), McpServer, and the four StandardPlugins.
- GraphStatistics counts widened to long.
- Backward compatible: existing databases read correctly (FilePath/StartLine were
  already stored in dedicated columns).

Fix: SitecoreUnicornPlugin set Relationship="REFERENCES" then RelationType=hint,
which (via the alias) silently overwrote Relationship with the hint. The hint is now
kept in the edge's Properties instead.

P3 UX
- Interactive Kanban: POST /api/interactions/status + GetNodeByIdAsync lets the
  dashboard change Task/Question/Decision/Milestone status; cards render per-type
  status buttons.
- Accessibility: aria-labels on icon-only buttons, aria-haspopup/aria-expanded on
  dropdown toggles, role=dialog on modals, keyboard-activatable language options,
  visible focus-visible outline.
- Error states: dashboard load/update surface toasts instead of silent console logs.

+++ src/Shonkor.Core/Models/GraphStatistics.cs
@@ -13,11 +13,11 @@ namespace Shonkor.Core.Models;
 /// (e.g., <c>Class</c>, <c>Method</c>, <c>File</c>).
 /// </param>
 /// <param name="EdgesByRelation">
-/// A breakdown of edge counts grouped by <see cref="GraphEdge.RelationType"/>
+/// A breakdown of edge counts grouped by <see cref="GraphEdge.Relationship"/>
 /// (e.g., <c>CONTAINS</c>, <c>IMPLEMENTS</c>, <c>CALLS</c>).
 /// </param>
 public record GraphStatistics(
-    int TotalNodes,
-    int TotalEdges,
+    long TotalNodes,
+    long TotalEdges,
     Dictionary<string, int> NodesByType,
     Dictionary<string, int> EdgesByRelation);
